+++ qCC/mainwindow.h
@@ -313,6 +313,8 @@ private slots:
 	
 	void doActionCreatePlane();
 	void doActionEditPlane();
+	void doActionFlipPlane();
+	void doActionComparePlanes();
 
 	void doActionDeleteScanGrids();
 	void doActionSmoothMeshSF();
@@ -333,6 +335,7 @@ private slots:
 	void doActionFlagMeshVertices();
 	void doActionSmoothMeshLaplacian();
 	void doActionSubdivideMesh();
+	void doActionFlipMeshTriangles();
 	void doActionComputeCPS();
 	void doActionShowWaveDialog();
 	void doActionCompressFWFData();